+++ modules/mesh-task/iam.tf
@@ -109,14 +109,14 @@ resource "aws_iam_policy" "execution" {
   "Version": "2012-10-17",
   "Statement": [
 %{if var.tls~}
-%{if var.consul_server_ca_cert_arn != ""~}
+%{if var.consul_ca_cert_arn != ""~}
     {
       "Effect": "Allow",
       "Action": [
         "secretsmanager:GetSecretValue"
       ],
       "Resource": [
-        "${var.consul_server_ca_cert_arn}"
+        "${var.consul_ca_cert_arn}"
       ]
     },
 %{endif~}
