+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -145,7 +145,7 @@ module "ecs_controller" {
   }
   launch_type                       = var.launch_type
   consul_bootstrap_token_secret_arn = module.consul_server.bootstrap_token_secret_arn
-  consul_server_address             = module.consul_server.server_dns
+  consul_server_hosts             = module.consul_server.server_dns
   consul_grpc_ca_cert_arn           = module.consul_server.ca_cert_arn
   consul_https_ca_cert_arn          = module.consul_server.ca_cert_arn
   ecs_cluster_arn                   = var.ecs_cluster_arn
@@ -223,7 +223,7 @@ EOT
       ]
     }
   ]
-  consul_server_address = module.consul_server.server_dns
+  consul_server_hosts = module.consul_server.server_dns
   upstreams = [
     {
       destinationName = "${var.server_service_name}_${var.suffix}"
@@ -275,7 +275,7 @@ module "test_server" {
     essential        = true
     logConfiguration = local.test_server_log_configuration
   }]
-  consul_server_address = module.consul_server.server_dns
+  consul_server_hosts = module.consul_server.server_dns
   log_configuration     = local.test_server_log_configuration
   port                  = 9090
