+++ test/acceptance/tests/validation/terraform/consul-ecs-config-validate/main.tf
@@ -5,7 +5,7 @@ provider "aws" {
   region = "us-west-2"
 }
 
-variable "upstreams_file" {
+variable "consul_ecs_config_file" {
   type = string
 }
 
@@ -15,7 +15,7 @@ module "test_client" {
   container_definitions = [{
     name = "basic"
   }]
-  consul_server_address = "consul.dc1"
+  consul_server_hosts = "consul.dc1.host"
   outbound_only         = true
-  upstreams             = jsondecode(file("${path.module}/${var.upstreams_file}"))
+  consul_ecs_config     = jsondecode(file("${path.module}/${var.consul_ecs_config_file}"))
 }
