+++ test/acceptance/tests/validation/terraform/consul-ecs-config-validate/test-complete-config.json
