+++ test/acceptance/tests/validation/terraform/consul-ecs-config-validate/test-empty-config.json
