+++ test/acceptance/tests/validation/terraform/consul-ecs-config-validate/test-invalid-config.json
