+++ test/acceptance/tests/validation/terraform/consul-ecs-config-validate/test-partial-config.json
