+++ test/acceptance/tests/validation/terraform/envoy-readiness-port-validate/main.tf
@@ -1,12 +1,11 @@
 # Copyright (c) HashiCorp, Inc.
 # SPDX-License-Identifier: MPL-2.0
-
 provider "aws" {
   region = "us-west-2"
 }
 
-variable "http_config_file" {
-  type = string
+variable "envoy_readiness_port" {
+  type = number
 }
 
 module "test_client" {
@@ -15,7 +14,7 @@ module "test_client" {
   container_definitions = [{
     name = "basic"
   }]
-  consul_server_address = "consul.dc1.host"
+  consul_server_hosts = "consul.dc1"
   outbound_only         = true
-  http_config       = jsondecode(file("${path.module}/${var.http_config_file}"))
+  envoy_readiness_port  = var.envoy_readiness_port
 }
