+++ test/acceptance/tests/validation/terraform/grpc-config-validate/test-complete-config.json
