+++ test/acceptance/tests/validation/terraform/grpc-config-validate/test-empty-config.json
