+++ test/acceptance/tests/validation/terraform/grpc-config-validate/test-invalid-config.json
