+++ test/acceptance/tests/validation/terraform/grpc-config-validate/test-partial-config.json
