+++ test/acceptance/tests/validation/terraform/http-config-validate/test-complete-config.json
