+++ test/acceptance/tests/validation/terraform/http-config-validate/test-empty-config.json
