+++ test/acceptance/tests/validation/terraform/http-config-validate/test-invalid-config.json
