+++ test/acceptance/tests/validation/terraform/http-config-validate/test-partial-config.json
