+++ test/acceptance/tests/validation/terraform/mesh-gateway-validate/main.tf
