+++ test/acceptance/tests/validation/terraform/public-listener-port-validate/main.tf
@@ -1,10 +1,11 @@
 # Copyright (c) HashiCorp, Inc.
 # SPDX-License-Identifier: MPL-2.0
+
 provider "aws" {
   region = "us-west-2"
 }
 
-variable "envoy_readiness_port" {
+variable "envoy_public_listener_port" {
   type = number
 }
 
@@ -14,7 +15,7 @@ module "test_client" {
   container_definitions = [{
     name = "basic"
   }]
-  consul_server_address = "consul.dc1"
-  outbound_only         = true
-  envoy_readiness_port  = var.envoy_readiness_port
+  consul_server_hosts      = "consul.dc1"
+  outbound_only              = true
+  envoy_public_listener_port = var.envoy_public_listener_port
 }
