+++ test/acceptance/tests/validation/terraform/service-name-validate/main.tf
@@ -5,7 +5,7 @@ provider "aws" {
   region = "us-west-2"
 }
 
-variable "grpc_config_file" {
+variable "consul_service_name" {
   type = string
 }
 
@@ -15,7 +15,7 @@ module "test_client" {
   container_definitions = [{
     name = "basic"
   }]
-  consul_server_address = "consul.dc1.host"
+  consul_server_hosts = "consul.dc1"
   outbound_only         = true
-  grpc_config       = jsondecode(file("${path.module}/${var.grpc_config_file}"))
+  consul_service_name   = var.consul_service_name
 }
