+++ test/acceptance/tests/validation/terraform/upstreams-validate/test-invalid-upstreams.json
