+++ test/acceptance/tests/validation/terraform/upstreams-validate/test-missing-localBindPort.json
