+++ test/acceptance/tests/validation/terraform/upstreams-validate/test-no-upstreams.json
