+++ test/acceptance/tests/validation/terraform/upstreams-validate/test-valid-upstreams.json
